# src/docs/swagger.yaml
basePath: /v1
definitions:
  httpapi.APIKeyCreateRequest:
    properties:
      name:
        type: string
      scope:
        type: string
    type: object
  httpapi.APIKeyCreateResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      scope:
        type: string
      token:
        type: string
      token_prefix:
        type: string
    type: object
  httpapi.APIKeyResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      revoked_at:
        type: string
      scope:
        type: string
      token_prefix:
        type: string
    type: object
  httpapi.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  httpapi.LoginResponse:
    properties:
      csrf_token:
        type: string
      session_expires_at:
        description: RFC3339
        type: string
    type: object
...
info:
  contact: {}
  description: sniply_api HTTP API.
  title: sniply_api API
  version: "1.0"
paths:
  /auth/api-keys:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/httpapi.APIKeyResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: List API keys
      tags:
      - auth
    post:
      consumes:
      - application/json
      parameters:
      - description: api key
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.APIKeyCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpapi.APIKeyCreateResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Create API key
      tags:
      - auth
  /auth/api-keys/{id}:
    delete:
      parameters:
      - description: api key id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Revoke API key
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.LoginResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Logout
      tags:
      - auth
...
securityDefinitions:
  ApiKeyAuth:
    description: 'API key (X-API-Key or Authorization: Bearer)'
    in: header
    name: X-API-Key
    type: apiKey
  SessionAuth:
    description: HttpOnly session cookie
    in: cookie
    name: sniply_session
    type: apiKey
swagger: "2.0"